refactor(auth): [LAT-22] restructuring files
